✨ - More response values
Closes #82.

Attach more information to the response. The response objects now include
status, statusText, ok, url, and headers. The headers object is a bit of an
anomoly as it's a Map on the client, but just a regular object in Node.
It seems like version 2 of node-fetch, will properly implement the headers
object in spec, but it is still in beta.